pyproject.toml
[project]
name = "telco-api-bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.18.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]
